--- duke.txt
@@ -1,5 +1,0 @@
-[T][X] read book 
-[D][ ] return book (by: Oct 5 2020 1800)
-[E][ ] project meeting  (from: Oct 2 2021 1200 to: Jan 5 2023 1300)
-[T][ ] join sports club 
-[T][ ] borrow book
