+++ src/main/java/DukeException.java
@@ -1,5 +1,3 @@
-package duke.exception;
-
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
