+++ src/main/java/ToDos.java
@@ -1,0 +1,12 @@
+public class ToDos extends Task{
+    public ToDos(String description) {
+        super(description);
+    }
+    public String getTaskIcon() {
+        return "T";
+    }
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
