+++ text-ui-test/input.txt
@@ -1,9 +1,8 @@
 todo read book
-deadline return book /by 2020-10-05 1800
-event project meeting /from 2021-10-02 1200 /to 2023-01-05 1300
+deadline return book /by June 6th
+event project meeting /from Aug 6th 2pm /to 4pm
 todo join sports club
 todo borrow book
 mark 1
 list
-find return book
 bye
